pyproject.toml
[tool.poetry]
name = "my-cloud-api"
version = "0.1.0"
description = ""
authors = ["admin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
django-rest-framework = "^0.1.0"
django-split-settings = "^1.3.2"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.27.2"
djoser = "^2.2.3"
django-filter = "^24.3"
drf-extra-fields = "^3.7.0"
filetype = "^1.2.0"

[tool.poetry.dev-dependencies]
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
